更新core_test文件
